refactor: clean up bootstrap layer
