+++ src/lib/bindInput.ts
@@ -1,0 +1,6 @@
+export function bindInput(
+  input: HTMLInputElement,
+  listener: (event: Event) => void,
+): void {
+  input.addEventListener("input", listener);
+}
