+++ src/main.ts
@@ -1,10 +1,10 @@
-import { getAndAssertHtmlElements } from "./lib/getAndAssertHtmlElements";
+import { getAppElements } from "./lib/getAppElements";
 import { setupApp } from "./lib/setupApp";
 import "./style.css";
 
 console.info("🚀 Running the app!");
 
-const elements = getAndAssertHtmlElements();
+const elements = getAppElements();
 const app = setupApp(elements);
 
 app.render();
